+++ src/editor/ui/mod.rs
@@ -22,12 +22,17 @@ pub use game_view::*;
 pub mod settings;
 pub use settings::*;
 
+pub mod tools;
+pub use tools::*;
+
 pub mod debug_panels;
 
 use bevy::{prelude::*, utils::HashMap, window::PrimaryWindow};
 use bevy_egui::{egui, EguiContext};
 
-use crate::{prelude::SelectedPlugin, EditorSet};
+use crate::{prelude::{SelectedPlugin, ToolExt}, EditorSet};
+
+use self::tools::gizmo::GizmoTool;
 
 use super::update_pan_orbit;
 
@@ -72,6 +77,9 @@ impl Plugin for EditorUiPlugin {
 
         app.add_plugins(SpaceHierarchyPlugin::default());
         app.add_plugins(SpaceInspectorPlugin);
+
+        app.editor_tool(GizmoTool::default());
+        app.world.resource_mut::<GameViewTab>().active_tool = Some(0);
 
         if self.use_standart_layout {
             let mut editor = app.world.resource_mut::<EditorUi>();
